+++ R/differential_expression.R
@@ -1,7 +1,7 @@
 #' @include seurat.R
 NULL
 
-globalVariables(names = 'avg_diff', package = 'Seurat', add = TRUE)
+globalVariables(names = 'avg_logFC', package = 'Seurat', add = TRUE)
 #' Gene expression markers of identity classes
 #'
 #' Finds markers (differentially expressed genes) for identity classes
@@ -11,28 +11,54 @@ globalVariables(names = 'avg_diff', package = 'Seurat', add = TRUE)
 #' @param ident.2 A second identity class for comparison. If NULL (default) -
 #' use all other cells for comparison.
 #' @param genes.use Genes to test. Default is to use all genes
-#' @param thresh.use Limit testing to genes which show, on average, at least
+#' @param logfc.threshold Limit testing to genes which show, on average, at least
 #' X-fold difference (log-scale) between the two groups of cells. Default is 0.25
-#' Increasing thresh.use speeds up the function, but can miss weaker signals.
-#' @param test.use Denotes which test to use. Seurat currently implements
-#' "bimod" (likelihood-ratio test for single cell gene expression, McDavid et
-#' al., Bioinformatics, 2013, default), "roc" (standard AUC classifier), "t"
-#' (Students t-test), and "tobit" (Tobit-test for differential gene expression,
-#' as in Trapnell et al., Nature Biotech, 2014), 'poisson', and 'negbinom'.
-#' The latter two options should only be used on UMI datasets, and assume an underlying
-#' poisson or negative-binomial distribution
-#' @param min.pct - only test genes that are detected in a minimum fraction of min.pct cells
-#' in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1
-#' @param min.diff.pct - only test genes that show a minimum difference in the fraction of detection between the two groups. Set to -Inf by default
+#' Increasing logfc.threshold speeds up the function, but can miss weaker signals.
+#' @param test.use Denotes which test to use. Available options are:
+##' \itemize{
+##'  \item{"wilcox"} : Wilcoxon rank sum test (default)
+##'  \item{"bimod"} : Likelihood-ratio test for single cell gene expression,
+##'  (McDavid et al., Bioinformatics, 2013)
+##'  \item{"roc"} : Standard AUC classifier
+##'  \item{"t"} : Student's t-test
+##'  \item{"tobit"} : Tobit-test for differential gene expression (Trapnell et
+##'  al., Nature Biotech, 2014)
+##'  \item{"poisson"} : Likelihood ratio test assuming an underlying poisson
+##'   distribution. Use only for UMI-based datasets
+##'  \item{"negbinom"} :  Likelihood ratio test assuming an underlying negative
+##'  binomial distribution. Use only for UMI-based datasets
+##'  \item{"MAST} : GLM-framework that treates cellular detection rate as a
+##'  covariate (Finak et al, Genome Biology, 2015)
+##'  \item{"DESeq2} : DE based on a model using the negative binomial
+##'  distribution (Love et al, Genome Biology, 2014)
+##' }
+#' @param min.pct  only test genes that are detected in a minimum fraction of
+#' min.pct cells in either of the two populations. Meant to speed up the function
+#' by not testing genes that are very infrequently expressed. Default is 0.1
+#' @param min.diff.pct  only test genes that show a minimum difference in the
+#' fraction of detection between the two groups. Set to -Inf by default
 #' @param only.pos Only return positive markers (FALSE by default)
-#' @param print.bar Print a progress bar once expression testing begins (uses pbapply to do this)
-#' @param max.cells.per.ident Down sample each identity class to a max number. Default is no downsampling. Not activated by default (set to Inf)
+#' @param print.bar Print a progress bar once expression testing begins (uses
+#' pbapply to do this)
+#' @param max.cells.per.ident Down sample each identity class to a max number.
+#' Default is no downsampling. Not activated by default (set to Inf)
 #' @param random.seed Random seed for downsampling
 #' @param latent.vars Variables to test
-#' @param min.cells Minimum number of cells expressing the gene in at least one of the two groups
-#'
-#' @return Matrix containing a ranked list of putative markers, and associated statistics (p-values, ROC score, etc.)
-#'
+#' @param min.cells Minimum number of cells expressing the gene in at least one
+#' of the two groups
+#' @param pseudocount.use Pseudocount to add to averaged expression values when
+#' calculating logFC. 1 by default.
+#' @param assay.type Type of assay to fetch data for (default is RNA)
+#' @param \dots Additional parameters to pass to specific DE functions
+#' @seealso \code{\link{MASTDETest}}, and \code{\link{DESeq2DETest}} for more information on these methods
+#' @return Matrix containing a ranked list of putative markers, and associated
+#' statistics (p-values, ROC score, etc.)
+#' @details p-value adjustment is performed using bonferroni correction based on
+#' the total number of genes in the dataset. Other correction methods are not
+#' recommended, as Seurat pre-filters genes using the arguments above, reducing
+#' the number of tests performed. Lastly, as Aaron Lun has pointed out, p-values
+#' should be interpreted cautiously, as the genes used for clustering are the
+#' same genes tested for differential expression.
 #' @import pbapply
 #'
 #' @export
@@ -46,8 +72,8 @@ FindMarkers <- function(
   ident.1,
   ident.2 = NULL,
   genes.use = NULL,
-  thresh.use = 0.25,
-  test.use = "bimod",
+  logfc.threshold = 0.25,
+  test.use = "wilcox",
   min.pct = 0.1,
   min.diff.pct = -Inf,
   print.bar = TRUE,
@@ -55,15 +81,18 @@ FindMarkers <- function(
   max.cells.per.ident = Inf,
   random.seed = 1,
   latent.vars = "nUMI",
-  min.cells = 3
+  min.cells = 3,
+  pseudocount.use = 1,
+  assay.type = "RNA",
+  ...
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
-  if (max.cells.per.ident < Inf) {
-    object <- SubsetData(
-      object = object,
-      max.cells.per.ident = max.cells.per.ident,
-      random.seed = random.seed
-    )
+  data.use <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.use))
+  methods.noprefiliter <- c("DESeq2", "zingeR")
+  if (test.use %in% methods.noprefiliter) {
+    genes.use <- rownames(x = data.use)
+    min.diff.pct <- -Inf
+    logfc.threshold <- 0
   }
   # in case the user passed in cells instead of identity classes
   if (length(x = as.vector(x = ident.1) > 1) && any(as.character(x = ident.1) %in% object@cell.names)) {
@@ -76,7 +105,8 @@ FindMarkers <- function(
     cells.2 <- intersect(x = ident.2, y = object@cell.names)
   } else {
     if (is.null(x = ident.2)) {
-      cells.2 <- object@cell.names
+      # cells.2 <- object@cell.names
+      cells.2 <- WhichCells(object = object,cells.use = setdiff(object@cell.names,cells.1))
     } else {
       cells.2 <- WhichCells(object = object, ident = ident.2)
     }
@@ -97,12 +127,11 @@ FindMarkers <- function(
   if (length(cells.2) < min.cells) {
     stop(paste("Cell group 2 has fewer than", as.character(min.cells), " cells in identity class", ident.2))
   }
-
   # gene selection (based on percent expressed)
   thresh.min <- 0
   data.temp1 <- round(
     x = apply(
-      X = object@data[genes.use, cells.1, drop = F],
+      X = data.use[genes.use, cells.1, drop = F],
       MARGIN = 1,
       FUN = function(x) {
         return(sum(x > thresh.min) / length(x = x))
@@ -113,7 +142,7 @@ FindMarkers <- function(
   )
   data.temp2 <- round(
     x = apply(
-      X = object@data[genes.use, cells.2, drop = F],
+      X = data.use[genes.use, cells.2, drop = F],
       MARGIN = 1,
       FUN = function(x) {
         return(sum(x > thresh.min) / length(x = x))
@@ -127,30 +156,37 @@ FindMarkers <- function(
   alpha.min <- apply(X = data.alpha, MARGIN = 1, FUN = max)
   names(x = alpha.min) <- rownames(x = data.alpha)
   genes.use <- names(x = which(x = alpha.min > min.pct))
-  if(length(genes.use) == 0) {
+  if (length(x = genes.use) == 0) {
     stop("No genes pass min.pct threshold")
   }
   alpha.diff <- alpha.min - apply(X = data.alpha, MARGIN = 1, FUN = min)
   genes.use <- names(
     x = which(x = alpha.min > min.pct & alpha.diff > min.diff.pct)
   )
-  if(length(genes.use) == 0) {
+  if (length(x = genes.use) == 0) {
     stop("No genes pass min.diff.pct threshold")
   }
 
   #gene selection (based on average difference)
-  data.1 <- apply(X = object@data[genes.use, cells.1, drop = F], MARGIN = 1, FUN = ExpMean)
-  data.2 <- apply(X = object@data[genes.use, cells.2, drop = F], MARGIN = 1, FUN = ExpMean)
+  data.1 <- apply(X = data.use[genes.use, cells.1, drop = F], MARGIN = 1, FUN = function(x) log(x = mean(x = expm1(x = x)) + pseudocount.use))
+  data.2 <- apply(X = data.use[genes.use, cells.2, drop = F], MARGIN = 1, FUN = function(x) log(x = mean(x = expm1(x = x)) + pseudocount.use))
   total.diff <- (data.1 - data.2)
-  genes.diff <- names(x = which(x = abs(x = total.diff) > thresh.use))
+  if (!only.pos) genes.diff <- names(x = which(x = abs(x = total.diff) > logfc.threshold))
+  if (only.pos) genes.diff <- names(x = which(x = total.diff > logfc.threshold))
   genes.use <- intersect(x = genes.use, y = genes.diff)
-  if(length(genes.use) == 0) {
-    stop("No genes pass thresh.use threshold")
+  if (length(x = genes.use) == 0) {
+    stop("No genes pass logfc.threshold threshold")
+  }
+  if (max.cells.per.ident < Inf) {
+    set.seed(seed = random.seed)
+    if (length(cells.1) > max.cells.per.ident) cells.1 = sample(x = cells.1, size = max.cells.per.ident)
+    if (length(cells.2) > max.cells.per.ident) cells.2 = sample(x = cells.2, size = max.cells.per.ident)
   }
   #perform DR
   if (test.use == "bimod") {
     to.return <- DiffExpTest(
       object = object,
+      assay.type = assay.type,
       cells.1 = cells.1,
       cells.2 = cells.2,
       genes.use = genes.use,
@@ -160,6 +196,7 @@ FindMarkers <- function(
   if (test.use == "roc") {
     to.return <- MarkerTest(
       object = object,
+      assay.type = assay.type,
       cells.1 = cells.1,
       cells.2 = cells.2,
       genes.use = genes.use,
@@ -169,6 +206,7 @@ FindMarkers <- function(
   if (test.use == "t") {
     to.return <- DiffTTest(
       object = object,
+      assay.type = assay.type,
       cells.1 = cells.1,
       cells.2 = cells.2,
       genes.use = genes.use,
@@ -178,6 +216,7 @@ FindMarkers <- function(
   if (test.use == "tobit") {
     to.return <- TobitTest(
       object = object,
+      assay.type = assay.type,
       cells.1 = cells.1,
       cells.2 = cells.2,
       genes.use = genes.use,
@@ -187,6 +226,7 @@ FindMarkers <- function(
   if (test.use == "negbinom") {
     to.return <- NegBinomDETest(
       object = object,
+      assay.type = assay.type,
       cells.1 = cells.1,
       cells.2 = cells.2,
       genes.use = genes.use,
@@ -198,6 +238,7 @@ FindMarkers <- function(
   if (test.use == "poisson") {
     to.return <- PoissonDETest(
       object = object,
+      assay.type = assay.type,
       cells.1 = cells.1,
       cells.2 = cells.2,
       genes.use = genes.use,
@@ -206,22 +247,59 @@ FindMarkers <- function(
       # min.cells # PoissonDETest doesn't have something for min.cells
     )
   }
+  if (test.use == "MAST") {
+    to.return <- MASTDETest(
+      object = object,
+      assay.type = assay.type,
+      cells.1 = cells.1,
+      cells.2 = cells.2,
+      genes.use = genes.use,
+      latent.vars = latent.vars,
+      ...
+    )
+  }
+
+  if (test.use == "wilcox") {
+    to.return <- WilcoxDETest(
+      object = object,
+      assay.type = assay.type,
+      cells.1 = cells.1,
+      cells.2 = cells.2,
+      genes.use = genes.use,
+      print.bar = print.bar,
+      ...
+    )
+  }
+    if (test.use == "DESeq2") {
+      to.return <- DESeq2DETest(
+        object = object,
+        assay.type = assay.type,
+        cells.1 = cells.1,
+        cells.2 = cells.2,
+        genes.use = genes.use,
+        ...
+      )
+  }
   #return results
-  to.return[, "avg_diff"] <- total.diff[rownames(x = to.return)]
+  to.return[, "avg_logFC"] <- total.diff[rownames(x = to.return)]
   to.return <- cbind(to.return, data.alpha[rownames(x = to.return), ])
+  to.return$p_val_adj = p.adjust(p = to.return$p_val,method = "bonferroni",
+                                 n =   nrow(GetAssayData(object = object,
+                                                         assay.type = assay.type,
+                                                         slot = "data")))
   if (test.use == "roc") {
-    to.return <- to.return[order(-to.return$power, -to.return$avg_diff), ]
+    to.return <- to.return[order(-to.return$power, -to.return$avg_logFC), ]
   } else {
-    to.return <- to.return[order(to.return$p_val, -to.return$avg_diff), ]
+    to.return <- to.return[order(to.return$p_val, -to.return$avg_logFC), ]
   }
   if (only.pos) {
-    to.return <- subset(x = to.return, subset = avg_diff > 0)
+    to.return <- subset(x = to.return, subset = avg_logFC > 0)
   }
   return(to.return)
 }
 
 globalVariables(
-  names = c('myAUC', 'p_val', 'avg_diff'),
+  names = c('myAUC', 'p_val', 'avg_logFC'),
   package = 'Seurat',
   add = TRUE
 )
@@ -229,22 +307,7 @@ globalVariables(
 #'
 #' Finds markers (differentially expressed genes) for each of the identity classes in a dataset
 #'
-#' @param object Seurat object
-#' @param genes.use Genes to test. Default is to all genes
-#' @param thresh.use Limit testing to genes which show, on average, at least
-#' X-fold difference (log-scale) between the two groups of cells.
-#' Increasing thresh.use speeds up the function, but can miss weaker signals.
-#' @param test.use Denotes which test to use. Seurat currently implements
-#' "bimod" (likelihood-ratio test for single cell gene expression, McDavid et
-#' al., Bioinformatics, 2013, default), "roc" (standard AUC classifier), "t"
-#' (Students t-test), and "tobit" (Tobit-test for differential gene expression,
-#' as in Trapnell et al., Nature Biotech, 2014), 'poisson', and 'negbinom'.
-#' The latter two options should only be used on UMI datasets, and assume an underlying
-#' poisson or negative-binomial distribution
-#' @param min.pct - only test genes that are detected in a minimum fraction of min.pct cells
-#' in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expression
-#' @param min.diff.pct - only test genes that show a minimum difference in the fraction of detection between the two groups. Set to -Inf by default
-#' @param only.pos Only return positive markers (FALSE by default)
+#' @inheritParams FindMarkers
 #' @param print.bar Print a progress bar once expression testing begins (uses pbapply to do this)
 #' @param max.cells.per.ident Down sample each identity class to a max number. Default is no downsampling.
 #' @param random.seed Random seed for downsampling
@@ -252,6 +315,8 @@ globalVariables(
 #' @param do.print FALSE by default. If TRUE, outputs updates on progress.
 #' @param min.cells Minimum number of cells expressing the gene in at least one of the two groups
 #' @param latent.vars remove the effects of these variables
+#' @param assay.type Type of assay to perform DE for (default is RNA)
+#' @param \dots Additional parameters to pass to specific DE functions
 #'
 #' @return Matrix containing a ranked list of putative markers, and associated
 #' statistics (p-values, ROC score, etc.)
@@ -264,10 +329,10 @@ globalVariables(
 FindAllMarkers <- function(
   object,
   genes.use = NULL,
-  thresh.use = 0.25,
+  logfc.threshold = 0.25,
   test.use = "bimod",
   min.pct = 0.1,
-  min.diff.pct = 0.05,
+  min.diff.pct = -Inf,
   print.bar = TRUE,
   only.pos = FALSE,
   max.cells.per.ident = Inf,
@@ -275,36 +340,42 @@ FindAllMarkers <- function(
   do.print = FALSE,
   random.seed = 1,
   min.cells = 3,
-  latent.vars = "nUMI"
+  latent.vars = "nUMI",
+  assay.type = "RNA",
+  ...
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
-  ident.use <- object@ident
+  data.1 <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.1))
   if ((test.use == "roc") && (return.thresh == 1e-2)) {
     return.thresh = 0.7
   }
   idents.all <- sort(x = unique(x = object@ident))
   genes.de <- list()
-  if (max.cells.per.ident < Inf) {
-    object <- SubsetData(
-      object = object,
-      max.cells.per.ident = max.cells.per.ident,
-      random.seed = random.seed
-    )
-  }
+  #if (max.cells.per.ident < Inf) {
+  #  object <- SubsetData(
+  #    object = object,
+  #    max.cells.per.ident = max.cells.per.ident,
+  #    random.seed = random.seed
+  #  )
+  #}
   for (i in 1:length(x = idents.all)) {
     genes.de[[i]] <- tryCatch(
       {
-        FindMarkers(object = object,
-                    ident.1 = idents.all[i],
-                    ident.2 = NULL,
-                    genes.use = genes.use,
-                    thresh.use = thresh.use,
-                    test.use = test.use,
-                    min.pct = min.pct,
-                    min.diff.pct = min.diff.pct,
-                    print.bar = print.bar,
-                    min.cells = min.cells,
-                    latent.vars = latent.vars
+        FindMarkers(
+          object = object,
+          assay.type = assay.type,
+          ident.1 = idents.all[i],
+          ident.2 = NULL,
+          genes.use = genes.use,
+          logfc.threshold = logfc.threshold,
+          test.use = test.use,
+          min.pct = min.pct,
+          min.diff.pct = min.diff.pct,
+          print.bar = print.bar,
+          min.cells = min.cells,
+          latent.vars = latent.vars,
+          max.cells.per.ident = max.cells.per.ident,
+          ...
         )
       },
       error = function(cond){
@@ -317,8 +388,10 @@ FindAllMarkers <- function(
   }
   gde.all <- data.frame()
   for (i in 1:length(x = idents.all)) {
+    if (is.null(x = unlist(x = genes.de[i]))) {
+      next
+    }
     gde <- genes.de[[i]]
-    if (is.null(unlist(gde))) next
     if (nrow(x = gde) > 0) {
       if (test.use == "roc") {
         gde <- subset(
@@ -326,7 +399,7 @@ FindAllMarkers <- function(
           subset = (myAUC > return.thresh | myAUC < (1 - return.thresh))
         )
       } else {
-        gde <- gde[order(gde$p_val, -gde$avg_diff), ]
+        gde <- gde[order(gde$p_val, -gde$avg_logFC), ]
         gde <- subset(x = gde, subset = p_val < return.thresh)
       }
       if (nrow(x = gde) > 0) {
@@ -339,8 +412,9 @@ FindAllMarkers <- function(
     }
   }
   if (only.pos) {
-    return(subset(x = gde.all, subset = avg_diff > 0))
+    return(subset(x = gde.all, subset = avg_logFC > 0))
   }
+  rownames(x = gde.all) <- make.unique(names = as.character(x = gde.all$gene))
   return(gde.all)
 }
 
@@ -353,6 +427,7 @@ FindAllMarkers <- function(
 #' @inheritParams FindMarkers
 #' @param node The node in the phylogenetic tree to use as a branch point
 #' @param tree.use Can optionally pass the tree to be used. Default uses the tree in object@@cluster.tree
+#' @param assay.type Type of assay to fetch data for (default is RNA)
 #' @param ... Additional arguments passed to FindMarkers
 #'
 #' @return Matrix containing a ranked list of putative markers, and associated
@@ -368,11 +443,16 @@ FindMarkersNode <- function(
   node,
   tree.use = NULL,
   genes.use = NULL,
-  thresh.use = 0.25,
+  logfc.threshold = 0.25,
   test.use = "bimod",
+  assay.type = "RNA",
   ...
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  data.use <- GetAssayData(
+    object = object,
+    assay.type = assay.type
+  )
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.use))
   tree <- SetIfNull(x = tree.use, default = object@cluster.tree[[1]])
   ident.order <- tree$tip.label
   nodes.1 <- ident.order[GetLeftDescendants(tree = tree, node = node)]
@@ -381,10 +461,11 @@ FindMarkersNode <- function(
   #print(nodes.2)
   to.return <- FindMarkers(
     object = object,
+    assay.type = assay.type,
     ident.1 = nodes.1,
     ident.2 = nodes.2,
     genes.use = genes.use,
-    thresh.use = thresh.use,
+    logfc.threshold = logfc.threshold,
     test.use = test.use,
     ...
   )
@@ -399,7 +480,7 @@ globalVariables(names = c('myAUC', 'p_val'), package = 'Seurat', add = TRUE)
 #' @param object Seurat object. Must have object@@cluster.tree slot filled. Use BuildClusterTree() if not.
 #' @param node Node from which to start identifying split markers, default is top node.
 #' @param genes.use Genes to test. Default is to use all genes
-#' @param thresh.use Limit testing to genes which show, on average, at least
+#' @param logfc.threshold Limit testing to genes which show, on average, at least
 #' X-fold difference (log-scale) between the two groups of cells.
 #' @param test.use Denotes which test to use. Seurat currently implements
 #' "bimod" (likelihood-ratio test for single cell gene expression, McDavid et
@@ -418,6 +499,8 @@ globalVariables(names = c('myAUC', 'p_val'), package = 'Seurat', add = TRUE)
 #' @param return.thresh Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)
 #' @param do.print Print status updates
 #' @param min.cells Minimum number of cells expressing the gene in at least one of the two groups
+#' @param assay.type Type of assay to fetch data for (default is RNA)
+#' @param \dots Additional parameters to pass to specific DE functions
 #'
 #' @return Returns a dataframe with a ranked list of putative markers for each node and associated statistics
 #'
@@ -426,13 +509,15 @@ globalVariables(names = c('myAUC', 'p_val'), package = 'Seurat', add = TRUE)
 #' @export
 #'
 #' @examples
+#' pbmc_small
+#'
 #' FindAllMarkersNode(pbmc_small)
 #'
 FindAllMarkersNode <- function(
   object,
   node = NULL,
   genes.use = NULL,
-  thresh.use = 0.25,
+  logfc.threshold = 0.25,
   test.use = "bimod",
   min.pct = 0.1,
   min.diff.pct = 0.05,
@@ -442,14 +527,16 @@ FindAllMarkersNode <- function(
   return.thresh = 1e-2,
   do.print = FALSE,
   random.seed = 1,
-  min.cells = 3
+  min.cells = 3,
+  assay.type = "RNA",
+  ...
 ) {
-  if(length(object@cluster.tree) == 0){
+  if (length(object@cluster.tree) == 0) {
     stop("Tree hasn't been built yet. Run BuildClusterTree to build.")
   }
-  genes.use <- SetIfNull(x = genes.use, default = rownames(object@data))
+  data.use <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.use))
   node <- SetIfNull(x = node, default = object@cluster.tree[[1]]$edge[1, 1])
-  ident.use <- object@ident
   tree.use <- object@cluster.tree[[1]]
   descendants <- DFT(tree = tree.use, node = node, path = NULL, include.children = TRUE)
   all.children <- sort(x = tree.use$edge[,2][!tree.use$edge[,2] %in% tree.use$edge[,1]])
@@ -459,17 +546,18 @@ FindAllMarkersNode <- function(
   orig.nodes <- c(node, as.numeric(setdiff(descendants, keep.children)))
   tree.use <- drop.tip(tree.use, drop.children)
   new.nodes <- unique(tree.use$edge[,1])
-  if ((test.use == 'roc') && (return.thresh==1e-2)) {
+  if ((test.use == 'roc') && (return.thresh == 1e-2)) {
     return.thresh <- 0.7
   }
   genes.de <- list()
-  for (i in ((tree.use$Nnode+2):max(tree.use$edge))) {
+  for (i in ((tree.use$Nnode + 2):max(tree.use$edge))) {
     genes.de[[i]] <- FindMarkersNode(
       object = object,
+      assay.type = assay.type,
       node = i,
       tree.use = tree.use,
       genes.use = genes.use,
-      thresh.use = thresh.use,
+      logfc.threshold = logfc.threshold,
       test.use = test.use,
       min.pct = min.pct,
       min.diff.pct = min.diff.pct,
@@ -484,11 +572,11 @@ FindAllMarkersNode <- function(
     }
   }
   gde.all <- data.frame()
-  for (i in ((tree.use$Nnode+2):max(tree.use$edge))) {
-    gde <- genes.de[[i]]
-    if (is.null(x = unlist(gde))) {
+  for (i in ((tree.use$Nnode + 2):max(tree.use$edge))) {
+    if (is.null(x = unlist(x = genes.de[i]))) {
       next
     }
+    gde <- genes.de[[i]]
     if (nrow(x = gde) > 0) {
       if (test.use == 'roc') {
         gde <- subset(
@@ -497,7 +585,7 @@ FindAllMarkersNode <- function(
         )
       }
       if ( (test.use == 'bimod') || (test.use == 't')) {
-        gde <- gde[order(gde$p_val,-gde$avg_diff), ]
+        gde <- gde[order(gde$p_val,-gde$avg_logFC), ]
         gde <- subset(x = gde, subset = p_val < return.thresh)
       }
       if (nrow(x = gde) > 0) {
@@ -509,9 +597,10 @@ FindAllMarkersNode <- function(
       }
     }
   }
-  gde.all$cluster <- MapVals(v = gde.all$cluster,
-                             from = new.nodes,
-                             to = orig.nodes
+  gde.all$cluster <- MapVals(
+    v = gde.all$cluster,
+    from = new.nodes,
+    to = orig.nodes
   )
   return(gde.all)
 }
@@ -523,6 +612,7 @@ FindAllMarkersNode <- function(
 #' @param ident.2 A second identity class for comparison. If NULL (default) - use all other cells
 #' for comparison.
 #' @param grouping.var grouping variable
+#' @param assay.type Type of assay to fetch data for (default is RNA)
 #' @param \dots parameters to pass to FindMarkers
 #'
 #' @return Matrix containing a ranked list of putative conserved markers, and associated statistics
@@ -546,6 +636,7 @@ FindConservedMarkers <- function(
   ident.1,
   ident.2 = NULL,
   grouping.var,
+  assay.type = "RNA",
   ...
 ) {
   object.var <- FetchData(object = object, vars.all = grouping.var)
@@ -598,6 +689,7 @@ FindConservedMarkers <- function(
     )
     marker.test[[i]] <- FindMarkers(
       object = object,
+      assay.type = assay.type,
       ident.1 = ident.use.1,
       ident.2 = ident.use.2,
       ...
@@ -641,6 +733,7 @@ FindConservedMarkers <- function(
 #' @param object Seurat object
 #' @param cells.1 Group 1 cells
 #' @param cells.2 Group 2 cells
+#' @param assay.type Type of assay to fetch data for (default is RNA)
 #' @return Returns a p-value ranked matrix of putative differentially expressed
 #' genes.
 #'
@@ -654,10 +747,12 @@ DiffExpTest <- function(
   object,
   cells.1,
   cells.2,
+  assay.type = "RNA",
   genes.use = NULL,
   print.bar = TRUE
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  data.test <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(data.test))
   if (print.bar) {
     iterate.fxn <- pblapply
   } else {
@@ -669,8 +764,8 @@ DiffExpTest <- function(
       FUN = function(x) {
         return(
           DifferentialLRT(
-            x = as.numeric(x = object@data[x, cells.1]),
-            y = as.numeric(x = object@data[x, cells.2])
+            x = as.numeric(x = data.test[x, cells.1]),
+            y = as.numeric(x = data.test[x, cells.2])
           )
         )
       }
@@ -692,6 +787,7 @@ DiffExpTest <- function(
 #' @param latent.vars Latent variables to test
 #' @param print.bar Print progress bar
 #' @param min.cells Minimum number of cells threshold
+#' @param assay.type Type of assay to fetch data for (default is RNA)
 #'
 #' @return Returns a p-value ranked matrix of putative differentially expressed
 #' genes.
@@ -715,18 +811,19 @@ NegBinomDETest <- function(
   genes.use = NULL,
   latent.vars = NULL,
   print.bar = TRUE,
-  min.cells = 3
+  min.cells = 3,
+  assay.type = "RNA"
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data")))
   # check that the gene made it through the any filtering that was done
-  genes.use <- genes.use[genes.use %in% rownames(x = object@data)]
+  genes.use <- genes.use[genes.use %in% rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data"))]
   my.latent <- FetchData(
     object = object,
     vars.all = latent.vars,
     cells.use = c(cells.1, cells.2),
     use.raw = TRUE
   )
-  to.test.data <- object@raw.data[genes.use, c(cells.1, cells.2)]
+  to.test.data <- GetAssayData(object = object,assay.type = assay.type,slot = "raw.data")[genes.use, c(cells.1, cells.2)]
   to.test <- data.frame(my.latent, row.names = c(cells.1, cells.2))
   to.test[cells.1, "group"] <- "A"
   to.test[cells.2, "group"] <- "B"
@@ -798,6 +895,7 @@ NegBinomDETest <- function(
 #' @param latent.vars Latent variables to test
 #' @param print.bar Print progress bar
 #' @param min.cells Minimum number of cells threshold
+#' @param assay.type Type of assay to fetch data for (default is RNA)
 #'
 #' @return Returns a p-value ranked data frame of test results.
 #'
@@ -821,11 +919,12 @@ NegBinomRegDETest <- function(
   genes.use = NULL,
   latent.vars = NULL,
   print.bar = TRUE,
-  min.cells = 3
+  min.cells = 3,
+  assay.type = "RNA"
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data")))
   # check that the gene made it through the any filtering that was done
-  genes.use <- genes.use[genes.use %in% rownames(x = object@data)]
+  genes.use <- genes.use[genes.use %in% rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data"))]
   print(
     sprintf(
       'NegBinomRegDETest for %d genes and %d and %d cells',
@@ -840,7 +939,7 @@ NegBinomRegDETest <- function(
       rep.int(x = 'B', times = length(x = cells.2))
     )
   )
-  to.test.data <- object@raw.data[genes.use, c(cells.1, cells.2), drop = FALSE]
+  to.test.data <- GetAssayData(object = object,assay.type = assay.type,slot = "raw.data")[genes.use, c(cells.1, cells.2), drop = FALSE]
   print('Calculating mean per gene per group')
   above.threshold <- pmax(
     apply(X = to.test.data[, cells.1] > 0, MARGIN = 1, FUN = mean),
@@ -914,6 +1013,7 @@ globalVariables(names = 'min.cells', package = 'Seurat', add = TRUE)
 #' @param genes.use Genes to use for test
 #' @param latent.vars Latent variables to test
 #' @param print.bar Print progress bar
+#' @param assay.type Type of assay to fetch data for (default is RNA)
 #'
 #' @return Returns a p-value ranked matrix of putative differentially expressed
 #' genes.
@@ -936,18 +1036,19 @@ PoissonDETest <- function(
   min.cells = 3,
   genes.use = NULL,
   latent.vars = NULL,
-  print.bar = TRUE
+  print.bar = TRUE,
+  assay.type = "RNA"
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data")))
   # check that the gene made it through the any filtering that was done
-  genes.use <- genes.use[genes.use %in% rownames(x = object@data)]
+  genes.use <- genes.use[genes.use %in% rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data"))]
   my.latent <- FetchData(
     object = object,
     vars.all = latent.vars,
     cells.use = c(cells.1, cells.2),
     use.raw = TRUE
   )
-  to.test.data <- object@raw.data[genes.use, c(cells.1, cells.2)]
+  to.test.data <- GetAssayData(object = object,assay.type = assay.type,slot = "raw.data")[genes.use, c(cells.1, cells.2)]
   to.test <- data.frame(my.latent, row.names = c(cells.1, cells.2))
   to.test[cells.1,"group"] <- "A"
   to.test[cells.2,"group"] <- "B"
@@ -1008,6 +1109,249 @@ PoissonDETest <- function(
   return(to.return)
 }
 
+#' Differential expression using MAST
+#'
+#' Identifies differentially expressed genes between two groups of cells using
+#' a hurdle model tailored to scRNA-seq data. Utilizes the MAST package to run
+#' the DE testing.
+#'
+#' @references Andrew McDavid, Greg Finak and Masanao Yajima (2017). MAST: Model-based
+#' Analysis of Single Cell Transcriptomics. R package version 1.2.1.
+#' https://github.com/RGLab/MAST/
+#' @param object Seurat object
+#' @param cells.1 Group 1 cells
+#' @param cells.2 Group 2 cells
+#' @param min.cells Minimum number of cells expressing the gene in at least one
+#' of the two groups
+#' @param genes.use Genes to use for test
+#' @param latent.vars Confounding variables to adjust for in DE test. Default is
+#' "nUMI", which adjusts for cellular depth (i.e. cellular detection rate). For
+#' non-UMI based data, set to nGene instead.
+#' @param assay.type Type of assay to fetch data for (default is RNA)
+#' @param \dots Additional parameters to zero-inflated regression (zlm) function
+#' in MAST
+#' @details
+#' To use this method, please install MAST, using instructions at https://github.com/RGLab/MAST/
+#'
+#' @return Returns a p-value ranked matrix of putative differentially expressed
+#' genes.
+#'
+#' @importFrom stats relevel
+#' @importFrom utils installed.packages
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#'   pbmc_small
+#'   MASTDETest(pbmc_small, cells.1 = WhichCells(object = pbmc_small, ident = 1),
+#'               cells.2 = WhichCells(object = pbmc_small, ident = 2))
+#' }
+#'
+MASTDETest <- function(
+  object,
+  cells.1,
+  cells.2,
+  min.cells = 3,
+  genes.use = NULL,
+  latent.vars = NULL,
+  assay.type = "RNA",
+  ...
+) {
+  # Check for MAST
+  if (!'MAST' %in% rownames(x = installed.packages())) {
+    stop("Please install MAST - learn more at https://github.com/RGLab/MAST")
+  }
+  data.test <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.test))
+  # check that the gene made it through the any filtering that was done
+  genes.use <- genes.use[genes.use %in% rownames(x = data.test)]
+  my.latent <- FetchData(
+    object = object,
+    vars.all = latent.vars,
+    cells.use = c(cells.1, cells.2),
+    use.raw = TRUE
+  )
+  if (length(x = latent.vars) > 0) {
+    my.latent <- scale(x = my.latent)
+  }
+  coldata <- data.frame(my.latent, row.names = c(cells.1, cells.2))
+  coldata[cells.1, "group"] <- "Group1"
+  coldata[cells.2, "group"] <- "Group2"
+  coldata$group <- factor(x = coldata$group)
+  coldata$wellKey <- rownames(x = coldata)
+  latent.vars <- c("condition", latent.vars)
+  countdata.test <- data.test[genes.use, rownames(x = coldata)]
+  fdat <- data.frame(rownames(x = countdata.test))
+  colnames(x = fdat)[1] <- "primerid"
+  rownames(x = fdat) <- fdat[, 1]
+  sca <- MAST::FromMatrix(
+    exprsArray = as.matrix(x = countdata.test),
+    cData = coldata,
+    fData = fdat
+  )
+  cond <- factor(x = SummarizedExperiment::colData(sca)$group)
+  cond <- relevel(x = cond, ref = "Group1")
+  SummarizedExperiment::colData(sca)$condition <- cond
+  fmla <- as.formula(
+    object = paste0(" ~ ", paste(latent.vars, collapse = "+"))
+  )
+  zlmCond <- MAST::zlm(formula = fmla, sca = sca, ...)
+  summaryCond <- summary(object = zlmCond, doLRT = 'conditionGroup2')
+  summaryDt <- summaryCond$datatable
+  # fcHurdle <- merge(
+  #   summaryDt[contrast=='conditionGroup2' & component=='H', .(primerid, `Pr(>Chisq)`)], #hurdle P values
+  #   summaryDt[contrast=='conditionGroup2' & component=='logFC', .(primerid, coef, ci.hi, ci.lo)], by='primerid'
+  # ) #logFC coefficients
+  # fcHurdle[,fdr:=p.adjust(`Pr(>Chisq)`, 'fdr')]
+  p_val <- summaryDt[summaryDt[, "component"] == "H", 4]
+  genes.return <- summaryDt[summaryDt[, "component"] == "H", 1]
+  # p_val <- subset(summaryDt, component == "H")[, 4]
+  # genes.return <- subset(summaryDt, component == "H")[, 1]
+  to.return <- data.frame(p_val, row.names = genes.return)
+  return(to.return)
+}
+
+#' Differential expression using DESeq2
+#'
+#' Identifies differentially expressed genes between two groups of cells using
+#' DESeq2
+#'
+#' @references Love MI, Huber W and Anders S (2014). "Moderated estimation of
+#' fold change and dispersion for RNA-seq data with DESeq2." Genome Biology.
+#' https://bioconductor.org/packages/release/bioc/html/DESeq2.html
+#' @param object Seurat object
+#' @param cells.1 Group 1 cells
+#' @param cells.2 Group 2 cells
+#' @param min.cells Minimum number of cells expressing the gene in at least one
+#' of the two groups
+#' @param genes.use Genes to use for test
+#' @param assay.type Type of assay to fetch data for (default is RNA)
+#' @param ... Extra parameters to pass to DESeq2::results
+#' @return Returns a p-value ranked matrix of putative differentially expressed
+#' genes.
+#'
+#' @details
+#' This test does not support pre-filtering of genes based on average difference
+#' (or percent detection rate) between cell groups. However, genes may be
+#' pre-filtered based on their minimum detection rate (min.pct) across both cell
+#' groups. To use this method, please install DESeq2, using the instructions at
+#'  https://bioconductor.org/packages/release/bioc/html/DESeq2.html
+#'
+#' @importFrom utils installed.packages
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#'   pbmc_small
+#'   DESeq2DETest(pbmc_small, cells.1 = WhichCells(object = pbmc_small, ident = 1),
+#'               cells.2 = WhichCells(object = pbmc_small, ident = 2))
+#' }
+#'
+DESeq2DETest <- function(
+  object,
+  cells.1,
+  cells.2,
+  min.cells = 3,
+  genes.use = NULL,
+  assay.type = "RNA",
+  ...
+) {
+  if (!'DESeq2' %in% rownames(x = installed.packages())) {
+    stop("Please install DESeq2 - learn more at https://bioconductor.org/packages/release/bioc/html/DESeq2.html")
+  }
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data")))
+  # check that the gene made it through the any filtering that was done
+  genes.use <- genes.use[genes.use %in% rownames(x = GetAssayData(object = object,assay.type = assay.type,slot = "data"))]
+  coldata <- object@meta.data[c(cells.1, cells.2), ]
+  coldata[cells.1, "group"] <- "Group1"
+  coldata[cells.2, "group"] <- "Group2"
+  coldata$group <- factor(x = coldata$group)
+  coldata$wellKey <- rownames(x = coldata)
+  countdata.test <- GetAssayData(object = object,assay.type = assay.type,slot = "raw.data")[genes.use, rownames(x = coldata)]
+  fdat <- data.frame(rownames(x = countdata.test))
+  colnames(x = fdat)[1] <- "primerid"
+  rownames(x = fdat) <- fdat[, 1]
+  dds1 <- DESeq2::DESeqDataSetFromMatrix(
+    countData = countdata.test,
+    colData = coldata,
+    design = ~ group
+  )
+  dds1 <- DESeq2::estimateSizeFactors(object = dds1)
+  dds1 <- DESeq2::estimateDispersions(object = dds1, fitType = "local")
+  dds1 <- DESeq2::nbinomWaldTest(object = dds1)
+  res <- DESeq2::results(
+    object = dds1,
+    contrast = c("group", "Group1", "Group2"),
+    alpha = 0.05,
+    ...
+  )
+  p_val <- res$pvalue
+  genes.return <- rownames(x = res)
+  to.return <- data.frame(p_val, row.names = genes.return)
+  return(to.return)
+}
+
+#' Differential expression using Wilcoxon Rank Sum
+#'
+#' Identifies differentially expressed genes between two groups of cells using
+#' a Wilcoxon Rank Sum test
+#' @param object Seurat object
+#' @param cells.1 Group 1 cells
+#' @param cells.2 Group 2 cells
+#' @param min.cells Minimum number of cells expressing the gene in at least one
+#' of the two groups
+#' @param genes.use Genes to use for test
+#' @param print.bar Print a progress bar
+#' @param assay.type Type of assay to perform DE for (default is RNA)
+#' @param ... Extra parameters passed to wilcox.test
+#'
+#' @return Returns a p-value ranked matrix of putative differentially expressed
+#' genes.
+#'
+#' @importFrom pbapply pbsapply
+#' @importFrom stats wilcox.test
+#'
+#' @export
+#'
+#' @examples
+#' pbmc_small
+#' WilcoxDETest(pbmc_small, cells.1 = WhichCells(object = pbmc_small, ident = 1),
+#'             cells.2 = WhichCells(object = pbmc_small, ident = 2))
+#'
+WilcoxDETest <- function(
+  object,
+  cells.1,
+  cells.2,
+  min.cells = 3,
+  genes.use = NULL,
+  print.bar = TRUE,
+  assay.type = "RNA",
+  ...
+) {
+  data.test <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.test))
+  # check that the gene made it through the any filtering that was done
+  genes.use <- genes.use[genes.use %in% rownames(x = data.test)]
+  coldata <- object@meta.data[c(cells.1, cells.2), ]
+  coldata[cells.1, "group"] <- "Group1"
+  coldata[cells.2, "group"] <- "Group2"
+  coldata$group <- factor(x = coldata$group)
+  coldata$wellKey <- rownames(x = coldata)
+  countdata.test <- data.test[genes.use, rownames(x = coldata)]
+  mysapply <- if (print.bar) {pbsapply} else {sapply}
+  p_val <- mysapply(
+    X = 1:nrow(x = countdata.test),
+    FUN = function(x) {
+      return(wilcox.test(countdata.test[x, ] ~ coldata$group, ...)$p.value)
+    }
+  )
+  genes.return <- rownames(x = countdata.test)
+  to.return <- data.frame(p_val, row.names = genes.return)
+  return(to.return)
+}
+
 #' Differential expression testing using Tobit models
 #'
 #' Identifies differentially expressed genes between two groups of cells using
@@ -1033,13 +1377,15 @@ TobitTest <- function(
   cells.1,
   cells.2,
   genes.use = NULL,
-  print.bar = TRUE
+  print.bar = TRUE,
+  assay.type = "RNA"
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  data.test <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.test))
   #print(genes.diff)
   to.return <- TobitDiffExpTest(
-    data1 = object@data[, cells.1],
-    data2 = object@data[, cells.2],
+    data1 = data.test[, cells.1],
+    data2 = data.test[, cells.2],
     mygenes = genes.use,
     print.bar = print.bar
   )
@@ -1078,12 +1424,14 @@ MarkerTest <- function(
   cells.1,
   cells.2,
   genes.use = NULL,
-  print.bar = TRUE
+  print.bar = TRUE,
+  assay.type = "RNA"
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
+  data.test <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.test))
   to.return <- AUCMarkerTest(
-    data1 = object@data[, cells.1],
-    data2 = object@data[, cells.2],
+    data1 = data.test[, cells.1],
+    data2 = data.test[, cells.2],
     mygenes = genes.use,
     print.bar = print.bar
   )
@@ -1117,10 +1465,12 @@ DiffTTest <- function(
   cells.1,
   cells.2,
   genes.use = NULL,
-  print.bar = TRUE
+  print.bar = TRUE,
+  assay.type = "RNA"
 ) {
-  genes.use <- SetIfNull(x = genes.use, default = rownames(x = object@data))
-  data.use <- object@data
+  data.test <- GetAssayData(object = object,assay.type = assay.type,slot = "data")
+  genes.use <- SetIfNull(x = genes.use, default = rownames(x = data.test))
+  # data.use <- object@data
   if (print.bar) {
     iterate.fxn=pblapply
   } else {
@@ -1130,7 +1480,7 @@ DiffTTest <- function(
     x = iterate.fxn(
       X = genes.use,
       FUN = function(x) {
-        t.test(x = object@data[x, cells.1], y = object@data[x, cells.2])$p.value
+        t.test(x = data.test[x, cells.1], y = data.test[x, cells.2])$p.value
       }
     )
   )