+++ NEWS.md
@@ -2,6 +2,14 @@
 All notable changes to Seurat will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [2.1.0] - 2017-10-12
+### Added
+- Support for using MAST and DESeq2 packages for differential expression testing in FindMarkers
+- Support for multi-modal single-cell data via @assay slot
+
+### Changed
+- Default DE test changed to Wilcoxon rank sum test
+
 ## [2.0.1] - 2017-08-18
 ### Added
  - Now available on CRAN